refactoring scripts project structure
