+++ godot_scripts/DeadZone.gdns
@@ -1,0 +1,5 @@
+[gd_resource type="NativeScript" format=2]
+
+[resource]
+resource_name = "DeadZone"
+class_name = "DeadZone"
