+++ godot_scripts/Player.gdns
@@ -1,7 +1,7 @@
 [gd_resource type="NativeScript" load_steps=2 format=2]
 
 [sub_resource type="GDNativeLibrary" id=1]
-entry/OSX.64 = "res://scripts/target/debug/libplayer.dylib"
+entry/OSX.64 = "res://scripts/target/debug/libscripts.dylib"
 entry/Windows.64 = "res://scripts/target/debug/player.dll"
 entry/Windows.32 = "res://scripts/target/debug/player.dll"
 entry/X11.64 = "res://scripts/target/debug/libplayer.so"
