+++ godot_scripts/Ui.gdns
@@ -1,7 +1,7 @@
 [gd_resource type="NativeScript" load_steps=2 format=2]
 
 [sub_resource type="GDNativeLibrary" id=1]
-entry/OSX.64 = "res://scripts/target/debug/libui.dylib"
+entry/OSX.64 = "res://scripts/target/debug/libscripts.dylib"
 dependency/OSX.64 = [  ]
 
 [resource]
