+++ scenes/Mundo.tscn
@@ -1,4 +1,4 @@
-[gd_scene load_steps=21 format=2]
+[gd_scene load_steps=23 format=2]
 
 [ext_resource path="res://assets/Tilesets/tileset_forest.png" type="Texture" id=1]
 [ext_resource path="res://scenes/Player.tscn" type="PackedScene" id=2]
@@ -10,6 +10,7 @@
 [ext_resource path="res://scenes/Coin.tscn" type="PackedScene" id=8]
 [ext_resource path="res://assets/Objects/coin_gold.png" type="Texture" id=9]
 [ext_resource path="res://godot_scripts/Ui.gdns" type="Script" id=10]
+[ext_resource path="res://godot_scripts/DeadZone.gdns" type="Script" id=11]
 
 [sub_resource type="ConvexPolygonShape2D" id=1]
 points = PoolVector2Array( 0, 0, 16, 0, 16, 16, 0, 16 )
@@ -324,6 +325,9 @@ corner_radius_top_right = 3
 corner_radius_bottom_right = 3
 corner_radius_bottom_left = 3
 
+[sub_resource type="RectangleShape2D" id=11]
+extents = Vector2( 194, 10 )
+
 [node name="Mundo" type="Node2D"]
 
 [node name="Sky" type="Sprite" parent="."]
@@ -390,4 +394,11 @@ margin_left = -100.0
 margin_top = -76.0
 margin_right = -83.0
 margin_bottom = -62.0
-text = "00"
+text = "0"
+
+[node name="Area2D" type="Area2D" parent="."]
+script = ExtResource( 11 )
+
+[node name="CollisionShape2D" type="CollisionShape2D" parent="Area2D"]
+position = Vector2( 7, 122 )
+shape = SubResource( 11 )
