+++ scripts/Cargo.toml
@@ -1,3 +1,11 @@
-[workspace] 
-members = [ "coin","player", "ui"]
-resolver = "2"
+[package]
+name = "scripts"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+gdnative = "0.11.0"
