--- scripts/coin/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "coin"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-gdnative = "0.11.0"
