--- scripts/coin/src/lib.rs
@@ -1,11 +1,0 @@
-use gdnative::prelude::*;
-
-mod coin;
-
-use coin::Coin;
-
-fn init(handle: InitHandle) {
-    handle.add_class::<Coin>();
-}
-
-godot_init!(init);
