--- scripts/player/src/lib.rs
@@ -1,10 +1,0 @@
-use gdnative::prelude::*;
-pub mod player;
-
-use player::Player;
-
-fn init(handle: InitHandle) {
-    handle.add_class::<Player>();
-}
-
-godot_init!(init);
