+++ scripts/src/classes/coin.rs
