+++ scripts/src/classes/mod.rs
@@ -1,0 +1,3 @@
+pub mod player;
+pub mod ui;
+pub mod coin;
