+++ scripts/src/classes/player.rs
