--- scripts/ui/src/lib.rs
@@ -1,11 +1,0 @@
-use gdnative::prelude::*;
-
-mod ui;
-
-use ui::Ui;
-
-fn init(handle: InitHandle) {
-    handle.add_class::<Ui>();
-}
-
-godot_init!(init);
